DEV_HANDLER: Fixed DEVICES instantiation. Compiles.

+++ dev_handler/devices.h
@@ -10,26 +10,28 @@ See hibikeDevices.json for the numbers. */
 
 /* The number of functional devices */
 #define NUM_DEVICES 11
-/* The largest device type number + 1
+
+/*
+ * The largest device type number + 1
  * DEVICES_LENGTH != NUM_DEVICES because some are NULL (Ex: type 6, 8, 9)
 */
 #define DEVICES_LENGTH 14
 
-/* Returns a pointer to the device for DEVICE_TYPE */
+/* Returns a pointer to the device given its DEVICE_TYPE */
 device_t* get_device(uint16_t device_type);
-/* Converts a device name to the 16-bit type. */
+/* Returns the device type given its device name DEV_NAME */
 uint16_t device_name_to_type(char* dev_name);
-/* Converts a device type to its name */
+/* Returns the name of a device given the DEV_TYPE. */
 char* device_type_to_name(uint16_t dev_type);
-/* Returns an array of param name strings given a device type */
+/* Fills PARAM_NAMES with all the names of the parameters for the given device type */
 void all_params_for_device_type(uint16_t dev_type, char* param_names[]);
-/* Check if PARAM of device type is readable */
-int readable(uint16_t dev_type, char* param_name);
-/* Check if PARAM of device type is writable */
-int writeable(uint16_t dev_type, char* param_name);
-/* Return the data type of the device type's parameter */
+/* Returns 1 if PARAM of device type is readable. Otherwise, 0. */
+uint8_t readable(uint16_t dev_type, char* param_name);
+/* Returns 1 if PARAM of device type is writable. Otherwise, 0. */
+uint8_t writeable(uint16_t dev_type, char* param_name);
+/* Return the data type of the device type's parameter ("int", "float", or "bool"). */
 char* get_param_type(uint16_t dev_type, char* param_name);
-/* Get the param id in array from param name */
+/* Get the param id of PARAM_NAME for the device DEV_TYPE */
 uint8_t get_param_id(uint16_t dev_type, char* param_name);
 
 #endif
